Add PhoneticDictionary as json file and refactor file structure
