+++ SpeechMod/Info.json
@@ -1,11 +1,11 @@
 {
   "Id": "SpeechMod",
-  "DisplayName": "SpeechMod - Text to Speech",
+  "DisplayName": "SpeechMod",
   "Author": "Osmodium",
   "ManagerVersion": "0.23.3",
   "Requirements": [],
   "AssemblyName": "SpeechMod.dll",
   "EntryMethod": "SpeechMod.Main.Load",
   "GameVersion": "1.0.4",
-  "Version": "0.9.4"
+  "Version": "0.9.5"
 }
