+++ SpeechMod/PhoneticDictionary.json
@@ -1,0 +1,9 @@
+{
+  "—": ",",
+  "Kenabres": "Keenaaabres",
+  "Iomedae": "I,omedaee",
+  "Golarion": "Goolaarion",
+  "Sovyrian": "Sovyyrian",
+  "Rovagug": "Rovaagug",
+  "Irabeth": "Iira,beth"
+}
