+++ SpeechMod/SpeechMod.csproj
@@ -37,6 +37,9 @@
     <Reference Include="Assembly-CSharp">
       <HintPath>E:\SteamLibrary\steamapps\common\Pathfinder Second Adventure\Wrath_Data\Managed\Assembly-CSharp.dll</HintPath>
     </Reference>
+    <Reference Include="Newtonsoft.Json, Version=13.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
+      <HintPath>..\packages\Newtonsoft.Json.13.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
+    </Reference>
     <Reference Include="Owlcat.Runtime.UI, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL">
       <SpecificVersion>False</SpecificVersion>
       <HintPath>E:\SteamLibrary\steamapps\common\Pathfinder Second Adventure\Wrath_Data\Managed\Owlcat.Runtime.UI.dll</HintPath>
@@ -66,18 +69,22 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ButtonFactory.cs" />
+    <Compile Include="Unity\ButtonFactory.cs" />
     <Compile Include="Patches\DialogCurrentPart_Patch.cs" />
     <Compile Include="Main.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Settings.cs" />
-    <Compile Include="Speech.cs" />
+    <Compile Include="Speech\Speech.cs" />
     <Compile Include="Patches\TooltipEngine_Patch.cs" />
-    <Compile Include="TextMeshProValues.cs" />
-    <Compile Include="WindowsVoiceUnity.cs" />
+    <Compile Include="Unity\TextMeshProValues.cs" />
+    <Compile Include="Unity\WindowsVoiceUnity.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Info.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Include="packages.config" />
+    <None Include="PhoneticDictionary.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
